Rework the wrapper to handle emulated environments
The wrapper can execute a toolchain program in a native or emulated
Poudriere jail, for aarch64, aarch64c, aarch64cb, riscv64 and riscv64c.
Depending on the case, an appropriate LD_LIBRARY_PATH variable must be
set to point it at toolchain libraries and base system libraries that
are either included in the native jail or mounted over nullfs from the
host.