Unbreak and clean up hacks to use host packages as hybrid packages
When we're building natively we want to make sure we don't use the
host's ABI, nor do we even want to use its pkg. It's quite possible I've
now broken something for the qemu-user case, but that should be fixed
following this approach, as the old approach was conceptually wrong at
its core.

This reinstates behaviour from prior to 5fdad4a3637f ("Choose a pkg
repository directory for pkg or pkg64"), e88ac9de2c19 ("Bootstrap hybrid
ABI pkg when starting a jail") and 3d222b145dd8 ("Install toolchain when
starting a jail") where native building worked just fine and didn't
pollute the jail with the host's ABI.

As part of this, we stop installing llvm-morello into /toolchain and put
it in /usr/local64 with all the other hybrid ABI packages (whether host
or actual hybrid). This means it lives where everyone expects it to for
the normal hybrid case, and the host "hybrid" case mirrors that rather
than unnecessarily diverging.

Also stop passing in various metadata about the jail and instead just
compute it inside the function, since it's always the same at every call
site and thus unnecessary boilerplate (and for the real hybrid case some
of the arguments are unused so them existing but not doing anything is
confusing).