Introduce the jail/bulk -H flag
The -H flag can be used to specify what local package repository for
replacement hybrid ABI packages should be used to build packages.

This flag is useful when a developer must test package builds with a
custom compiler. In such a case, they can create a CheriBSD ports branch
with a modified compiler port, build the replacement hybrid ABI packages
on a build host and then build packages using the local package
repository with the modified compiler.